requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2
json2html==1.3.0